Move serde string macros to internals
The macros are internal things and can live in `internals`. This will
help with future crate smashing.
